Implement Access Control on Users' Post CRUD

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -71,6 +81,13 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        $user_id = auth()->user()->id;
+
+        // Check access control
+        if ($post->user_id !== $user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -105,6 +122,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        $user_id = auth()->user()->id;
+
+        // Check access control
+        if ($post->user_id !== $user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page.');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post deleted.');
